Pontus-Execution-Layer/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aioredis==2.0.1
asyncpg==0.29.0
# Routing Engine Dependencies
ortools==9.9.3963
numpy==1.24.3
# Production Dependencies
slowapi==0.1.9
python-multipart==0.0.6
# CPLEX is optional - see ROUTING_ENGINE_SETUP.md for installation

